pyproject.toml
[project]
name = "coding-time-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arrow>=1.3.0",
    "google-api-python-client>=2.164.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "loguru>=0.7.3",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pystray>=0.19.5",
    "pywin32>=309",
]

[tool.ruff]
line-length = 80
lint.select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # isort
    "UP",   # pyupgrade
    "C4",   # flake8-comprehensions
    "F",    # pyflakes
]
lint.ignore = []

[tool.ruff.lint.isort]
known-first-party = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
